Milestone: Add Deployment  (Responsive design)
## Summary of PR

In this pull request, I implemented the following project requirements:

Deploy your website using GitHub Pages.
Check the online version of your portfolio and make sure that the page works properly.
Update the README of your repository to include a link to the online version.
If by any chance you have deployed your app earlier, open an issue with a link to the online version of your app. Otherwise, submit a link to the PR with changes required for the deployment.